feat: parser_utils.cppにメソッドの定義を移動
