+++ srcs/configuration/parser_utils.hpp
@@ -7,19 +7,8 @@
 
 namespace ParserUtils {
 
-    std::vector<std::string> extractTokensUntilSemicolon(std::vector<std::string>& tokens) {
-        std::vector<std::string> extracted_tokens;
-
-        while (!tokens.empty()) {
-            std::string token = tokens.front();
-            tokens.erase(tokens.begin());
-            if (token == ";") {
-                break;
-            }
-            extracted_tokens.push_back(token);
-        }
-        return extracted_tokens;
-    }
+    std::vector<std::string> extractTokensFromBlock(std::vector<std::string>& tokens);
+    std::vector<std::string> extractTokensUntilSemicolon(std::vector<std::string>& tokens);
 
 }
