+++ srcs/configuration/server_directive.cpp
@@ -1,5 +1,4 @@
 #include "server_directive.hpp"
-#include "parser_utils.hpp"
 
 ServerDirective::ServerDirective() {}
